dAIly_digest/requirements.txt
# Daily Digest Dependencies
# Core libraries
requests>=2.31.0
python-dotenv>=1.0.0

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# OpenAI
openai>=1.3.0

# Email handling (built-in libraries, but listing for completeness)
# smtplib - built-in
# email - built-in

# Date/time handling (built-in)
# datetime - built-in

# JSON handling (built-in)
# json - built-in

# File operations (built-in)
# os - built-in

# Logging (built-in)
# logging - built-in

# Type hints (built-in in Python 3.5+)
# typing - built-in

# Random selection (built-in)
# random - built-in

# System functions (built-in)
# sys - built-in

# Time utilities (built-in)
# time - built-in 